+++ examples/dreamcast/basic/fpu/exc/fpu_exc.c
@@ -24,7 +24,7 @@ int main(int argc, char **argv) {
     cont_btn_callback(0, CONT_START, (cont_btn_callback_t)arch_exit);
 
     fpscr = _get_fpscr();
-    printf("fpscr is %08x\n", fpscr);
+    printf("fpscr is %08x\n", (unsigned int)fpscr);
 
     while(d > 0.0) {
         printf("%f\n", d);
